+++ R/Gibbs_OMIFA.R
@@ -85,7 +85,7 @@
     Qs             <- Qs[index]
     psi.inv        <- psi.inv[,index, drop=FALSE]
     z              <- factor(z, labels=match(nn.ind, index))
-    z              <- as.numeric(levels(z))[z]
+    z              <- as.integer(levels(z))[z]
     G.non          <- G
     if(burnin       < 1)  {
       mu.store[,,1]           <- mu
@@ -120,7 +120,7 @@
       Qs             <- Qs[index]
       psi.inv        <- psi.inv[,index, drop=FALSE]
       z              <- factor(z, labels=match(nn.ind, index))
-      z              <- as.numeric(levels(z))[z]
+      z              <- as.integer(levels(z))[z]
 
     # Cluster Labels
       psi            <- 1/psi.inv
@@ -268,7 +268,7 @@
       }
     }
     if(verbose)         close(pb)
-    Gmax             <- seq_len(max(as.numeric(z.store)))
+    Gmax             <- seq_len(max(as.integer(z.store)))
     Qmax             <- seq_len(max(Q.store))
     mu.store         <- if(sw["mu.sw"])  tryCatch(mu.store[,Gmax,, drop=FALSE],        error=function(e) mu.store)
     eta.store        <- if(sw["s.sw"])   tryCatch(eta.store[,Qmax,, drop=FALSE],       error=function(e) eta.store)
